+++ src/frontend-common/ini_settings_interface.h
@@ -13,7 +13,7 @@ public:
   INISettingsInterface(std::string filename);
   ~INISettingsInterface() override;
 
-  //bool Save() override;
+  bool Save() override;
 
   void Clear() override;
 
@@ -30,7 +30,7 @@ public:
   //void ClearSection(const char* section) override;
 
   std::vector<std::string> GetStringList(const char* section, const char* key) override;
-  //void SetStringList(const char* section, const char* key, const std::vector<std::string>& items) override;
+  void SetStringList(const char* section, const char* key, const std::vector<std::string>& items) override;
   bool RemoveFromStringList(const char* section, const char* key, const char* item) override;
   bool AddToStringList(const char* section, const char* key, const char* item) override;